Moved the fetch call to the highets level so that it only has to be made once.
