+++ src/components/SearchBar.js
@@ -1,14 +1,17 @@
 import React from 'react'
-import {Link} from 'react-router-dom'
+import { Link } from 'react-router-dom'
 
 const SearchBar = (props) => {
-    return(
-        <div className='searchBar'>
-           <input></input>
-        
-        </div>
-    )
-   
+  return (
+    <div className='searchBar'>
+        <form>
+          <input type="text" />
+          <input type="submit" value="submit" />
+        </form>
+
+    </div>
+  )
+
 }
 
 export default SearchBar
