+++ src/components/TopTen.js
@@ -1,27 +1,14 @@
 import React, { useState, useEffect } from 'react';
+import App from '../App.js';
 
-const TopTen = () => {
-  const [movies, setMovies] = useState([]);
+const TopTen = ({data}) => { 
 
-  useEffect(() => {
-    const fetchMovies = async () => {
-      try {
-        const response = await fetch('http://localhost:4000/movies/');
-        const data = await response.json();
-        setMovies(data.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt)).slice(0, 10));
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    fetchMovies();
-  }, []);
 
   return (
     <div>
       <h2>Top 10/ Most Recent</h2>
       <ul>
-        {movies.map(movie => (
+        {data.map(movie => (
           <li key={movie._id}>
             <h3>{movie.title}</h3>
             <img src={movie.img} alt={movie.title} />
@@ -30,6 +17,6 @@ const TopTen = () => {
       </ul>
     </div>
   );
-};
+        };
 
 export default TopTen;
