--- src/pages/Home.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import TopTen from '../components/TopTen';
-
-const Home = () => {
-  return (
-    <div>
-      <h1>This is the home page</h1>
-      <TopTen />
-    </div>
-  );
-};
-
-export default Home;
